+++ lib/media-mgmt/dailyMediaManager.ts
@@ -3,14 +3,22 @@ import { MediaStreamRecorder, WavStreamPlayer } from "../wavtools";
 
 import Daily, {
   DailyCall,
+  DailyCameraErrorObject,
+  DailyCameraErrorType,
   DailyEventObjectAvailableDevicesUpdated,
+  DailyEventObjectCameraError,
   DailyEventObjectLocalAudioLevel,
   DailyEventObjectSelectedDevicesUpdated,
   DailyEventObjectTrack,
   DailyParticipant,
   DailyParticipantsObject,
 } from "@daily-co/daily-js";
-import { Participant, Tracks } from "@pipecat-ai/client-js";
+import {
+  DeviceArray,
+  DeviceError,
+  Participant,
+  Tracks,
+} from "@pipecat-ai/client-js";
 
 export class DailyMediaManager extends MediaManager {
   private _daily: DailyCall;
@@ -74,6 +82,7 @@ export class DailyMediaManager extends MediaManager {
       "selected-devices-updated",
       this._handleSelectedDevicesUpdated.bind(this),
     );
+    this._daily.on("camera-error", this.handleDeviceError.bind(this));
     this._daily.on("local-audio-level", this._handleLocalAudioLevel.bind(this));
   }
 
@@ -307,6 +316,59 @@ export class DailyMediaManager extends MediaManager {
     }
   }
 
+  private handleDeviceError(ev: DailyEventObjectCameraError) {
+    const generateDeviceError = (
+      error: DailyCameraErrorObject<DailyCameraErrorType>,
+    ) => {
+      const devices: DeviceArray = [];
+      switch (error.type) {
+        case "permissions": {
+          error.blockedMedia.forEach((d) => {
+            devices.push(d === "video" ? "cam" : "mic");
+          });
+          return new DeviceError(devices, error.type, error.msg, {
+            blockedBy: error.blockedBy,
+          });
+        }
+        case "not-found": {
+          error.missingMedia.forEach((d) => {
+            devices.push(d === "video" ? "cam" : "mic");
+          });
+          return new DeviceError(devices, error.type, error.msg);
+        }
+        case "constraints": {
+          error.failedMedia.forEach((d) => {
+            devices.push(d === "video" ? "cam" : "mic");
+          });
+          return new DeviceError(devices, error.type, error.msg, {
+            reason: error.reason,
+          });
+        }
+        case "cam-in-use": {
+          devices.push("cam");
+          return new DeviceError(devices, "in-use", error.msg);
+        }
+        case "mic-in-use": {
+          devices.push("mic");
+          return new DeviceError(devices, "in-use", error.msg);
+        }
+        case "cam-mic-in-use": {
+          devices.push("cam");
+          devices.push("mic");
+          return new DeviceError(devices, "in-use", error.msg);
+        }
+        case "undefined-mediadevices":
+        case "unknown":
+        default: {
+          devices.push("cam");
+          devices.push("mic");
+          return new DeviceError(devices, error.type, error.msg);
+        }
+      }
+    };
+    this._callbacks.onDeviceError?.(generateDeviceError(ev.error));
+  }
+
   private _handleLocalAudioLevel(ev: DailyEventObjectLocalAudioLevel) {
     this._callbacks.onLocalAudioLevel?.(ev.audioLevel);
   }
@@ -322,13 +384,17 @@ export class DailyMediaManager extends MediaManager {
         const status = this._mediaStreamRecorder.getStatus();
         switch (status) {
           case "ended":
-            await this._mediaStreamRecorder.begin(event.track);
-            if (this._connected) {
-              this._startRecording();
-              if (this._connectResolve) {
-                this._connectResolve();
-                this._connectResolve = null;
+            try {
+              await this._mediaStreamRecorder.begin(event.track);
+              if (this._connected) {
+                this._startRecording();
+                if (this._connectResolve) {
+                  this._connectResolve();
+                  this._connectResolve = null;
+                }
               }
+            } catch (e) {
+              // void. nothing to do.
             }
             break;
           case "paused":
@@ -338,8 +404,12 @@ export class DailyMediaManager extends MediaManager {
           default:
             if (this._currentAudioTrack !== event.track) {
               await this._mediaStreamRecorder.end();
-              await this._mediaStreamRecorder.begin(event.track);
-              this._startRecording();
+              try {
+                await this._mediaStreamRecorder.begin(event.track);
+                this._startRecording();
+              } catch (e) {
+                // void. nothing to do.
+              }
             } else {
               console.warn(
                 "track-started event received for current track and already recording",