+++ transports/daily/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to **Pipecat Daily WebRTC Transport** will be documented in 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+- Add support for generating `onDeviceError` callbacks/events when receiving a camera or speaker error.
+
 ## [1.0.0]
 
 - PipecatClient and Transport 1.0.0 Updates: